requirements.txt
# Core ML/NLP dependencies
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
joblib>=1.3.0,<2.0.0

# Dataset loading
datasets>=2.14.0,<3.0.0
huggingface-hub>=0.19.0,<1.0.0

# LLM API
groq>=0.4.0,<1.0.0
openai>=1.0.0,<2.0.0  # Optional, for OpenAI API support

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
rich>=13.0.0,<14.0.0

# Configuration
python-dotenv>=1.0.0,<2.0.0

# Development tools (optional)
pytest>=7.4.0,<8.0.0
black>=23.0.0,<24.0.0
flake8>=6.0.0,<7.0.0
mypy>=1.5.0,<2.0.0
pre-commit>=3.3.0,<4.0.0

# Documentation (optional)
sphinx>=7.0.0,<8.0.0
sphinx-rtd-theme>=1.3.0,<2.0.0 